Bug fix in lmbench show_result and updated stats script
1) lmbench show_result function in stats script
always prints the result of the first test result
file which is run with lkp because it reads result
from $(testbox).0 file which should be deleted with
each test run. The path to delete this file is
currently wrong in existing script. In this patch
it is updated with correct path.
2) Updated the stats script to print the MMAP read
latency.
3) Renamed largest_bandwidth function to generic name
largest_value to get the max latency as well as
bandwidth.

Signed-off-by: Swapnil Sapkal <swapnil.sapkal@amd.com>